move common menus to own file
